requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
azure-ai-language-conversations==1.1.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-search-documents==11.5.2
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.7
etelemetry==0.3.1
faiss-cpu==1.9.0
filelock==3.16.1
fitz==0.0.1.dev2
Flask==3.0.3
frontend==0.0.3
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.18
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2
langsmith==0.1.143
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mpmath==1.3.0
msrest==0.7.1
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.0
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.0.0
plotly==5.24.1
propcache==0.2.0
prov==2.0.1
puremagic==1.28
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydot==3.0.2
PyMuPDF==1.24.13
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytils==0.4.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.0
setuptools==75.5.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.41.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
tools==0.1.9
torch==2.5.1
tqdm==4.67.0
traits==6.4.3
transformers==4.46.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
Werkzeug==3.0.6
yarl==1.17.1
zipp==3.21.0
